Wow. This is actually pretty cool. Built a canvas display object that cycles through frames of simulation (constructed at run time). Then all the objects get updated over time in their own loops.
It's a pretty decent solution to not simulating everything, it just costs extra ram to run this whole setup. That will have to come down in the longer term.

Also, had to do a small hack for Chrome -- which seems to cache svg results unless you make small updates to the width (or something) every frame.

Right now, need the ability to investigate a single object -- and then publish that object. That's the missing component -- everything else has been written. Then I need the ability to do a phylogeny thing of iesor individuals.

They may not walk far yet, but soon by god they will...