Another nice update. Added separate socket server for just processing simulations. There is now publish simulations, when you press on an individual, the publish UI appears, and you can watch the object in real time. Closes and stops simulation on publish/clear click. Need another button: evolve that starts a new evolution with the chosen object.
Bad news-- firefox is pretty much a bust. Therefore, it's probably best to ensure this is wrapped by chrome or webkit for visuals to work.

Later experimentation might reduce the firefox lag issues -- but now is not the time.

I will need to try it in qt browser as well -- no ideas on how it will run.

Next step: publish button hookup to local server, and the ability to start from the wid in the html query params.