The form of win-IESoR at alife submission.
There are many sync bugs to fix, but those are usually indicative of system-wide architecture choices. I believe the easiest way to solve them is to readjust the evolution to be less driven by an internal uncontrollable loop. Not sure exactly how to do this, but the current priority is getting win-Picbreeder and winark setup.

Then testing/modifying win-iesor.